千逐数字资产/吧唧生成器服务端 python + vue/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
ReportLab==4.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
PyMySQL==1.1.0
cryptography==41.0.7
psutil==5.9.6
